+++ src/Racmacs_types.h
@@ -275,6 +275,7 @@ SEXP wrap(const AcMap& acmap){
   // Assemable list
   List out = List::create(
     _["name"] = acmap.name,
+    _["description"] = acmap.description,
     _["dilution_stepsize"] = acmap.dilution_stepsize,
     _["antigens"] = antigens,
     _["sera"] = sera,
@@ -648,6 +649,7 @@ AcMap as(SEXP sxp){
 
   // Attributes
   if(list.containsElementNamed("name")) acmap.name = as<std::string>(list["name"]);
+  if(list.containsElementNamed("description")) acmap.description = as<std::string>(list["description"]);
   if(list.containsElementNamed("dilution_stepsize")) acmap.dilution_stepsize = as<double>(list["dilution_stepsize"]);
   if(list.containsElementNamed("pt_drawing_order")){
     acmap.set_pt_drawing_order(