+++ tests/testthat/test-map_methods.R
@@ -6,17 +6,33 @@ context("Test reading and editing of chart details")
 # Load the map and the chart
 map <- read.acmap(filename = test_path("../testdata/testmap.ace"))
 
+test_that("Edit map details", {
+
+  # Initial default checks
+  expect_equal(mapName(map), "")
+  expect_equal(mapDescription(map), "")
+
+  # Making the edits
+  edited_map <- map
+  mapName(edited_map) <- "NEW NAME"
+  mapDescription(edited_map) <- "A map description"
+
+  # Checking the edits
+  expect_equal(mapName(edited_map), "NEW NAME")
+  expect_equal(mapDescription(edited_map), "A map description")
+
+  # Saving and loading a map
+  tmp <- tempfile(fileext = ".ace")
+  save.acmap(edited_map, tmp)
+  loaded_map <- read.acmap(tmp)
+
+  # Checking the loaded map
+  expect_equal(mapName(loaded_map), "NEW NAME")
+  expect_equal(mapDescription(loaded_map), "A map description")
+
+})
+
 test_that("Edit map strain details", {
-
-  # Chart name -------
-  expect_equal(
-    mapName(map),
-    ""
-  )
-
-  mapName(map) <- "NEW NAME"
-  expect_equal(mapName(map), "NEW NAME")
-
 
   # HI table -------
   expect_equal(dim(titerTable(map)), c(10, 5))