+++ NEWS.md
@@ -85,3 +85,4 @@
 * Dimension testing now calculates proportion of titers to exclude as test set as proportion of measured titers, not proportion of detectable titers.
 * Fixes an error when only one blob would be displayed in the viewer in cases where more than one blob was associated with a point
 * Speed improvements to the calculation of contour blobs from point densities in 2-dimensions
+* Added the option to set a description field for the map with the `mapDescription()` function